commit spring 3
Interface para funcionalidades comunes, ABM

+++ src/main/java/com/vet2C/vet_2C/Service/DuenioService.java
@@ -2,6 +2,7 @@ package com.vet2C.vet_2C.Service;
 
 import com.vet2C.vet_2C.Entity.Duenio;
 import com.vet2C.vet_2C.Repository.DuenioRepository;
+import com.vet2C.vet_2C.Service.Base.InterfaceService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,28 +12,40 @@ import java.util.Optional;
 
 @Service
 @AllArgsConstructor(onConstructor = @__(@Autowired))
-public class DuenioService {
+public class DuenioService implements InterfaceService<Duenio> {
 
     private final DuenioRepository duenioRepository;
-
-
-    public Duenio registrarDuenio(Duenio duenio){
+    @Override
+    public Duenio registrarEntidad(Duenio duenio) {
         return duenioRepository.save(duenio);
     }
 
-    public Optional<Duenio> buscarPorId(Long id){
+    @Override
+    public Optional<Duenio> buscarPorId(Long id) {
         return duenioRepository.findById(id);
     }
 
-    public List<Duenio> listarTodos(){
+    @Override
+    public Duenio modificarEntidad(Duenio duenio) {
+        return duenioRepository.save(duenio);
+    }
+
+    @Override
+    public void eliminarEntidad(Long id) {
+        Optional<Duenio> duenio = duenioRepository.findById(id);
+        if(duenio.isPresent())
+        {
+            duenioRepository.deleteById(id);
+        }
+    }
+
+    @Override
+    public Optional<Duenio> buscarEntidadPorString(String s) {
+        return duenioRepository.findByNombreIgnoreCase(s);
+    }
+
+    @Override
+    public List<Duenio> listarEntidades() {
         return duenioRepository.findAll();
-    }
-
-    public void eliminarDuenio(Long id){
-        duenioRepository.deleteById(id);
-    }
-
-    public Optional<Duenio> buscarPornombre(String nombre){
-        return duenioRepository.findByNombre(nombre);
     }
 }
